Add check-post-exists functionality
